+++ docs/changelog.rst
@@ -68,6 +68,30 @@ A release with a number of small and medium fixes, with no specific focus.
   supported by Mopidy is Python 2.7. We're continuously working towards running
   Mopidy on Python 3. (Fixes: :issue:`344`)
 
+**Core**
+
+- Tracklist has now the power to make decisions on which is the next/previous
+  song, along with previously playback associated features, such as randomness,
+  consumption, repeat and single. For that, a new method has been created to
+  mark songs, some Playback properties have been converted into functions and
+  both functions and properties have been moved into Tracklist to have more
+  modularity:
+
+  - Properties merged into functions: :attr:`tracklist_position` merged to
+    :meth:`index`
+
+  - Properties moved: :attr:`random`, :attr:`repeat`, :attr:`consume` and
+    :attr:`single`
+
+  - Method created from properties: :meth:`next_track` from 
+    `tl_track_at_next`, :meth:`eot_track` from `tl_track_at_eot` and
+    :meth:`previous_track` from `tl_track_at_previous`
+
+  - Method created to separe functionality: :meth:`mark` 
+
+- Tracklist's get_tl_track_at_* and tracklist_position now need a tl_track as a
+  reference to give an answer.
+
 **Command line options**
 
 - Converted from the optparse to the argparse library for handling command line